+++ src/compas_rhino/geometry/surface.py
@@ -31,46 +31,83 @@ class RhinoSurface(BaseRhinoGeometry):
         guid = compas_rhino.select_surface()
         return cls.from_guid(guid)
 
-    def to_compas(self, cls=None):
+    def to_compas(self, cls=None, facefilter=None, cleanup=True):
         """Convert the surface b-rep loops to a COMPAS mesh.
 
         Parameters
         ----------
         cls : :class:`compas.datastructures.Mesh`, optional
             The type of COMPAS mesh.
+        facefilter : callable, optional
+            A filter for selection which Brep faces to include.
+            If provided, the filter should return ``True`` or ``False`` per face.
+            A very simple filter that includes all faces is ``def facefilter(face): return True``.
+            Default parameter value is ``None`` in which case all faces are included.
+        cleanup : bool, optional
+            Flag indicating to clean up the result.
+            Cleaning up means to remove isolated faces and unused vertices.
+            Default is ``True``.
 
         Returns
         -------
         :class:`compas.datastructures.Mesh`
             The resulting mesh.
 
+        Examples
+        --------
+        >>> import compas_rhino
+        >>> from compas_rhino.geometry import RhinoSurface
+        >>> from compas_rhino.artists import MeshArtist
+
+        >>> def facefilter(face):
+        ...     success, w, h = face.GetSurfaceSize()
+        ...     if success:
+        ...         if w > 10 and h > 10:
+        ...             return True
+        ...     return False
+        ...
+
+        >>> guid = compas_rhino.select_surface()
+        >>> surf = RhinoSurface.from_guid(guid)
+        >>> mesh = surf.to_compas(facefilter=facefilter)
+
+        >>> artist = MeshArtist(mesh, layer="Blocks")
+        >>> artist.clear_layer()
+        >>> artist.draw()
+
         """
         if not self.geometry.HasBrepForm:
             return
         brep = Rhino.Geometry.Brep.TryConvertBrep(self.geometry)
+        if facefilter and callable(facefilter):
+            brepfaces = [face for face in brep.Faces if facefilter(face)]
+        else:
+            brepfaces = brep.Faces
+        # vertex maps and face lists
         gkey_xyz = {}
         faces = []
-        for loop in brep.Loops:
+        for face in brepfaces:
+            loop = face.OuterLoop
             curve = loop.To3dCurve()
             segments = curve.Explode()
-            face = []
-            sp = segments[0].PointAtStart
-            ep = segments[0].PointAtEnd
-            sp_gkey = geometric_key(sp)
-            ep_gkey = geometric_key(ep)
-            gkey_xyz[sp_gkey] = sp
-            gkey_xyz[ep_gkey] = ep
-            face.append(sp_gkey)
-            face.append(ep_gkey)
+            a = segments[0].PointAtStart
+            b = segments[0].PointAtEnd
+            a_gkey = geometric_key(a)
+            b_gkey = geometric_key(b)
+            gkey_xyz[a_gkey] = a
+            gkey_xyz[b_gkey] = b
+            face = [a_gkey, b_gkey]
             for segment in segments[1:-1]:
-                ep = segment.PointAtEnd
-                ep_gkey = geometric_key(ep)
-                face.append(ep_gkey)
-                gkey_xyz[ep_gkey] = ep
+                b = segment.PointAtEnd
+                b_gkey = geometric_key(b)
+                face.append(b_gkey)
+                gkey_xyz[b_gkey] = b
             faces.append(face)
+        # vertices and faces
         gkey_index = {gkey: index for index, gkey in enumerate(gkey_xyz)}
         vertices = [list(xyz) for gkey, xyz in gkey_xyz.items()]
-        faces = [[gkey_index[gkey] for gkey in f] for f in faces]
+        faces = [[gkey_index[gkey] for gkey in face] for face in faces]
+        # remove duplicates from vertexlist
         polygons = []
         for temp in faces:
             face = []
@@ -78,63 +115,18 @@ class RhinoSurface(BaseRhinoGeometry):
                 if vertex not in face:
                     face.append(vertex)
             polygons.append(face)
+        # define mesh type
         cls = cls or Mesh
+        # create mesh
         mesh = cls.from_vertices_and_faces(vertices, polygons)
         mesh.name = self.name
+        # remove isolated faces
+        if cleanup:
+            for face in list(mesh.faces()):
+                if not mesh.face_neighbors(face):
+                    mesh.delete_face(face)
+            mesh.remove_unused_vertices()
         return mesh
-
-    # def uv_to_compas(self, cls=None, density=(10, 10)):
-    #     """Convert the surface UV space to a COMPAS mesh.
-
-    #     Parameters
-    #     ----------
-    #     cls : :class:`compas.datastructures.Mesh`, optional
-    #         The type of mesh.
-    #     density : tuple of int, optional
-    #         The density in the U and V directions.
-    #         Default is ``u = 10`` and ``v = 10``.
-
-    #     Returns
-    #     -------
-    #     :class:`compas.datastructures.Mesh`
-    #         The COMPAS mesh.
-    #     """
-    #     return self.heightfield_to_compas(cls=cls, density=density, over_space=True)
-
-    # def heightfield_to_compas(self, cls=None, density=(10, 10), over_space=False):
-    #     """Convert a heightfiled of the surface to a COMPAS mesh.
-
-    #     Parameters
-    #     ----------
-    #     cls : :class:`compas.datastructures.Mesh`, optional
-    #         The type of mesh.
-    #     density : tuple of int, optional
-    #         The density in the two grid directions.
-    #         Default is ``u = 10`` and ``v = 10``.
-    #     over_space : bool, optional
-    #         Construct teh grid over the surface UV space instead of the XY axes.
-    #         Default is ``False``.
-
-    #     Returns
-    #     -------
-    #     :class:`compas.datastructures.Mesh`
-    #         The COMPAS mesh.
-    #     """
-    #     try:
-    #         u, v = density
-    #     except Exception:
-    #         u, v = density, density
-    #     vertices = self.heightfield(density=(u, v), over_space=over_space)
-    #     faces = []
-    #     for i in range(u - 1):
-    #         for j in range(v - 1):
-    #             face = [(i + 0) * v + j,
-    #                     (i + 1) * v + j,
-    #                     (i + 1) * v + j + 1,
-    #                     (i + 0) * v + j + 1]
-    #             faces.append(face)
-    #     cls = cls or Mesh
-    #     return cls.from_vertices_and_faces(vertices, faces)
 
     # ==========================================================================
     #